MTDSA-27408: Align RetrieveUkPropertyAnnualSubmission packages
