+++ test/v5/retrieveUkPropertyAnnualSubmission/RetrieveUkPropertyAnnualSubmissionConnectorSpec.scala
@@ -23,6 +23,7 @@ import api.models.outcomes.ResponseWrapper
 import org.scalamock.handlers.CallHandler
 import v5.retrieveUkPropertyAnnualSubmission.RetrieveUkPropertyAnnualSubmissionConnector._
 import v5.retrieveUkPropertyAnnualSubmission.def1.model.request.Def1_RetrieveUkPropertyAnnualSubmissionRequestData
+import v5.retrieveUkPropertyAnnualSubmission.def1.model.response.Def1_RetrieveUkPropertyAnnualSubmissionResponse
 import v5.retrieveUkPropertyAnnualSubmission.def1.model.response.ukFhlProperty.RetrieveUkFhlProperty
 import v5.retrieveUkPropertyAnnualSubmission.def1.model.response.ukProperty.RetrieveUkProperty
 import v5.retrieveUkPropertyAnnualSubmission.model.response._
