+++ test/v5/retrieveUkPropertyAnnualSubmission/def2/response/ukProperty/RetrieveUkPropertyStructuredBuildingAllowanceSpec.scala
@@ -18,7 +18,7 @@ package v5.retrieveUkPropertyAnnualSubmission.def2.response.ukProperty
 
 import play.api.libs.json.{JsValue, Json}
 import support.UnitSpec
-import v5.retrieveUkPropertyAnnualSubmission.def2.model.response.ukProperty._
+import v5.retrieveUkPropertyAnnualSubmission.def2.model.response._
 
 class RetrieveUkPropertyStructuredBuildingAllowanceSpec extends UnitSpec {
